+++ src/typings/typings.d.ts
@@ -1,5 +1,0 @@
-import { WindowType } from '../types/window';
-
-declare global {
-  interface Window extends WindowType {}
-}