+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-### v3.0.0
+# v4.0.0
+- [+] новые миксины: `mvk`, `odr`, `scaleWidthHeight`, `aspectRatio`, `fixSafariRadiusOverflow`, `customScrollbar`, `ifFlexGapNotSupported`, `flexGap`
+- [+] новые функции: `px-to-rem`, `round`, `fluid`, `safe-top-value`, `safe-bottom-value`
+- [*] в миксинах `backgroundImageCover`, `backgroundImageContain` параметр `image` стал необязательным
+- [*] миксин `centerPos` теперь принимает в качестве параметра объект с настройками: выбор оси - `axis`, дополнительные значения transform - `properties` и возможность использовать transform3d вместо transform - `is-3d`
+- [*] утилита markup переписана на классах, появилась возможность обращаться к `currentHtmlFontSize`
+- [-] из зависимостей убрана библиотека @ktsstudio/mediaproject-utils
+
+# v3.0.0
 - [-] удалены миксины и утилиты, использующие переменные vkui
 
 ### v2.0.3
@@ -13,13 +21,13 @@
 - [*] markup defaultMobileSize изменен на размер iphone x
 - [*] опечатка в миксине centerPosY (closing issue #4)
 
-## v2.0.0
+# v2.0.0
 - [+] appearAnimation
 - [+] утилиты для работы с отступами и хедером в ВК-миниаппе
 - [*] исправлены миксины с использованием глобальных классов
 - [*] fadeAnimation принимает длительность анимации и параметр наличия анимации пульсации
 
-## v1.0.0
+# v1.0.0
 - [+] checkIOS
 - [+] checkMobile
 - [+] markup